+++ lib/redis_client/resp3.rb
@@ -58,32 +58,6 @@ class RedisClient
       String.new(encoding: Encoding::BINARY, capacity: 128)
     end
 
-    def coerce_command!(command)
-      command = command.flat_map do |element|
-        case element
-        when Hash
-          element.flatten
-        when Set
-          element.to_a
-        else
-          element
-        end
-      end
-
-      command.map! do |element|
-        case element
-        when String
-          element
-        when Integer, Float, Symbol
-          element.to_s
-        else
-          raise TypeError, "Unsupported command argument type: #{element.class}"
-        end
-      end
-
-      command
-    end
-
     def dump_any(object, buffer)
       method = DUMP_TYPES.fetch(object.class) do
         raise TypeError, "Unsupported command argument type: #{object.class}"