+++ lib/redis_client/pooled.rb
@@ -49,40 +49,48 @@ class RedisClient
       pool.size
     end
 
-    %w(pipelined).each do |method|
-      class_eval <<~RUBY, __FILE__, __LINE__ + 1
-        def #{method}(&block)
-          with { |r| r.#{method}(&block) }
-        end
-      RUBY
-    end
-
-    %w(multi).each do |method|
-      class_eval <<~RUBY, __FILE__, __LINE__ + 1
-        def #{method}(**kwargs, &block)
-          with { |r| r.#{method}(**kwargs, &block) }
-        end
-      RUBY
-    end
-
-    %w(call call_once blocking_call pubsub).each do |method|
-      class_eval <<~RUBY, __FILE__, __LINE__ + 1
-        def #{method}(*args)
-          with { |r| r.#{method}(*args) }
-        end
-      RUBY
-    end
-
-    %w(scan sscan hscan zscan).each do |method|
-      class_eval <<~RUBY, __FILE__, __LINE__ + 1
-        def #{method}(*args, &block)
-          unless block_given?
-            return to_enum(__callee__, *args)
+    methods = %w(pipelined multi pubsub call call_once blocking_call)
+    iterable_methods = %w(scan sscan hscan zscan)
+    begin
+      methods.each do |method|
+        class_eval <<~RUBY, __FILE__, __LINE__ + 1
+          def #{method}(...)
+            with { |r| r.#{method}(...) }
           end
+        RUBY
+      end
 
-          with { |r| r.#{method}(*args, &block) }
-        end
-      RUBY
+      iterable_methods.each do |method|
+        class_eval <<~RUBY, __FILE__, __LINE__ + 1
+          def #{method}(...)
+            unless block_given?
+              return to_enum(__callee__, ...)
+            end
+
+            with { |r| r.#{method}(...) }
+          end
+        RUBY
+      end
+    rescue SyntaxError
+      methods.each do |method|
+        class_eval <<~RUBY, __FILE__, __LINE__ + 1
+          def #{method}(*args, &block)
+            with { |r| r.#{method}(*args, &block) }
+          end
+        RUBY
+      end
+
+      iterable_methods.each do |method|
+        class_eval <<~RUBY, __FILE__, __LINE__ + 1
+          def #{method}(*args, &block)
+            unless block_given?
+              return to_enum(__callee__, *args)
+            end
+
+            with { |r| r.#{method}(*args, &block) }
+          end
+        RUBY
+      end
     end
 
     private